Added support for kill statistics of a world that doesn't exist

+++ tibiakt-core/src/main/kotlin/com/galarzaa/tibiakt/core/parsers/KillStatisticsParser.kt
@@ -1,3 +1,19 @@
+/*
+ * Copyright © 2022 Allan Galarza
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
 package com.galarzaa.tibiakt.core.parsers
 
 
@@ -10,22 +26,22 @@ import com.galarzaa.tibiakt.core.utils.formData
 import com.galarzaa.tibiakt.core.utils.offsetStart
 import com.galarzaa.tibiakt.core.utils.parseTablesMap
 import com.galarzaa.tibiakt.core.utils.rows
-import org.jsoup.Jsoup
-import org.jsoup.nodes.Document
 import org.jsoup.nodes.Element
 
-object KillStatisticsParser : Parser<KillStatistics> {
-    override fun fromContent(content: String): KillStatistics {
-        val document: Document = Jsoup.parse(content)
-        val boxContent =
-            document.selectFirst("div.BoxContent") ?: throw ParsingException("BoxContent container not found")
+object KillStatisticsParser : Parser<KillStatistics?> {
+    override fun fromContent(content: String): KillStatistics? {
+        val boxContent = boxContent(content)
+        val formData =
+            boxContent.selectFirst("form")?.formData() ?: throw ParsingException("Could not find world selection form")
+        if (formData.values["world"].isNullOrBlank())
+            return null
         val tables = boxContent.parseTablesMap("table.Table1, table.Table3")
         return killStatistics {
             tables["Kill Statistics"]?.apply {
                 parseKillStatisticsTable(this)
             } ?: throw ParsingException("kill statistics table not found")
             val form = boxContent.selectFirst("form") ?: throw ParsingException("could not find form in value")
-            world = form.formData().values["world"] ?: throw ParsingException("could not find world value in form")
+            world = formData.values["world"] ?: throw ParsingException("could not find world value in form")
         }
     }
